+++ packages/browser-wallet/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 -   An issue where changing the credential metadata URL to an invalid URL, or a URL that does not contain a credential metadata file, would result in an empty screen.
+-   The wallet now ensures that the verifiable credential index used when adding a credential has not already been used in the contract.
 
 ## 1.1.3